finish

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: yhadari <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/10 17:28:59 by yhadari           #+#    #+#             */
-/*   Updated: 2019/10/17 16:25:32 by yhadari          ###   ########.fr       */
+/*   Updated: 2019/10/28 17:20:14 by yhadari          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i] != '\0')
